fix: moving things around
