+++ esbuild.js
@@ -7,9 +7,7 @@
 import { build } from 'esbuild'
 import path from 'path'
 
-await build({
-  banner: {
-    js: `#!/usr/bin/env node
+const banner = `#!/usr/bin/env node
 /*
  * Copyright IBM Corp. 2023, 2023
  *
@@ -19,11 +17,29 @@ await build({
 const require = (await import("node:module")).createRequire(import.meta.url);
 const __filename = (await import("node:url")).fileURLToPath(import.meta.url);
 const __dirname = (await import("node:path")).dirname(__filename);`
+
+await build({
+  banner: {
+    js: banner
   },
   bundle: true,
-  entryPoints: [path.join('dist', 'main', 'index.js'), path.join('dist', 'main', 'collect.js')],
+  entryPoints: [path.join('dist', 'main', 'collect.js')],
   format: 'esm',
-  outdir: 'dist',
+  minify: true,
+  outfile: path.join('dist', 'collect.js'),
+  platform: 'node',
+  target: 'es2021'
+})
+
+await build({
+  banner: {
+    js: banner
+  },
+  bundle: true,
+  entryPoints: [path.join('dist', 'main', 'background-process.js')],
+  format: 'esm',
+  minify: true,
+  outfile: path.join('dist', 'background-process.js'),
   platform: 'node',
   target: 'es2021'
 })
